[IMP] udes_core: add picking summary to form view
Extend stock picking form view to also show a summary of useful
information at the list of stock moves page. Also add ordered quantity
colunt to the list of stock moves and highligh discrepancies so that
it is easier for the users to check the state of the picking.

User-story: 1035